Packaged files and renamed Duke to InvMgr
