+++ src/main/java/seedu/duke/InvMgr.java
@@ -2,13 +2,17 @@ package seedu.duke;
 
 import seedu.duke.commands.ByeCommand;
 import seedu.duke.commands.Command;
+import seedu.duke.data.ItemList;
+import seedu.duke.parser.Parser;
+import seedu.duke.storage.Storage;
+import seedu.duke.ui.Ui;
 
-public class Duke {
+public class InvMgr {
     private Storage storage;
     private ItemList itemList;
     private Ui ui;
 
-    public Duke(String filePath) {
+    public InvMgr(String filePath) {
         ui = new Ui();
         storage = new Storage(filePath);
         itemList = new ItemList(storage.load());
@@ -27,9 +31,9 @@ public class Duke {
     }
 
     /**
-     * Main entry-point for the java.duke.Duke application.
+     * Main entry-point for the java.duke.InvMgr application.
      */
     public static void main(String[] args) {
-        new Duke("data/inventoryData.json").run();
+        new InvMgr("data/inventoryData.json").run();
     }
 }
