+++ src/main/java/seedu/duke/commands/ByeCommand.java
@@ -1,7 +1,7 @@
 package seedu.duke.commands;
 
-import seedu.duke.ItemList;
-import seedu.duke.Ui;
+import seedu.duke.data.ItemList;
+import seedu.duke.ui.Ui;
 
 public class ByeCommand extends Command {
     public void execute(Ui ui, ItemList itemList) {
