+++ src/main/java/seedu/duke/commands/Command.java
@@ -1,7 +1,7 @@
 package seedu.duke.commands;
 
-import seedu.duke.ItemList;
-import seedu.duke.Ui;
+import seedu.duke.data.ItemList;
+import seedu.duke.ui.Ui;
 
 public abstract class Command {
     public abstract void execute(Ui ui, ItemList itemList);
