+++ src/main/java/seedu/duke/commands/HiCommand.java
@@ -1,7 +1,7 @@
 package seedu.duke.commands;
 
-import seedu.duke.ItemList;
-import seedu.duke.Ui;
+import seedu.duke.data.ItemList;
+import seedu.duke.ui.Ui;
 
 public class HiCommand extends Command {
     public void execute(Ui ui, ItemList itemList) {
