+++ src/main/java/seedu/duke/storage/Storage.java
@@ -1,4 +1,6 @@
-package seedu.duke;
+package seedu.duke.storage;
+
+import seedu.duke.data.Item;
 
 import java.util.ArrayList;
